Update the tests for image size limits to test the current limits and series
Also fix the size limits returned for Ubuntu Server to be integers instead
of floats, and adjust the test output so that on error you can tell which
image is giving the error.

+++ lib/cdimage/tree.py
@@ -1880,13 +1880,13 @@ class DailyTreePublisher(Publisher):
             if self.config["DIST"] >= "impish":
                 # Requested by paride in #ubuntu-release
                 # https://irclogs.ubuntu.com/2021/08/31/%23ubuntu-release.html#t16:22
-                return 1.4 * 1000 * 1000 * 1000
+                return int(1.4 * 1000 * 1000 * 1000)
             elif self.config["DIST"] >= "focal":
                 # Size bump due to HWE kernel inclusion in the images
-                return 1.4 * 1000 * 1000 * 1000
+                return int(1.4 * 1000 * 1000 * 1000)
             else:
                 # email with powersj, 20200108
-                return 1.2 * 1000 * 1000 * 1000
+                return int(1.2 * 1000 * 1000 * 1000)
         else:
             if self.publish_type == "dvd":
                 # http://en.wikipedia.org/wiki/DVD_plus_RW
